chore(integrations): move django to internal
