+++ ddtrace/contrib/django/__init__.py
@@ -184,9 +184,13 @@ required_modules = ["django"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from . import patch as _patch
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.django.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.django.patch import patch as _patch
+        from ..internal.django.patch import get_version
+        from ..internal.django.patch import patch
+        from ..internal.django.patch import unpatch
 
         __all__ = ["patch", "unpatch", "_patch", "get_version"]
