+++ ddtrace/contrib/django/restframework.py
@@ -1,33 +1,15 @@
-import rest_framework.views
+from ddtrace.internal.utils.deprecations import DDTraceDeprecationWarning
+from ddtrace.vendor.debtcollector import deprecate
 
-from ddtrace.vendor.wrapt import wrap_function_wrapper as wrap
-
-from ..trace_utils import iswrapped
-from ..trace_utils import with_traced_module
+from ..internal.django.restframework import *  # noqa: F401,F403
 
 
-@with_traced_module
-def _traced_handle_exception(django, pin, wrapped, instance, args, kwargs):
-    """Sets the error message, error type and exception stack trace to the current span
-    before calling the original exception handler.
-    """
-    span = pin.tracer.current_span()
+def __getattr__(name):
+    deprecate(
+        ("%s.%s is deprecated" % (__name__, name)),
+        category=DDTraceDeprecationWarning,
+    )
 
-    if span is not None:
-        span.set_traceback()
-
-    return wrapped(*args, **kwargs)
-
-
-def patch_restframework(django):
-    """Patches rest_framework app.
-
-    To trace exceptions occurring during view processing we currently use a TraceExceptionMiddleware.
-    However the rest_framework handles exceptions before they come to our middleware.
-    So we need to manually patch the rest_framework exception handler
-    to set the exception stack trace in the current span.
-    """
-
-    # trace the handle_exception method
-    if not iswrapped(rest_framework.views.APIView, "handle_exception"):
-        wrap("rest_framework.views", "APIView.handle_exception", _traced_handle_exception(django))
+    if name in globals():
+        return globals()[name]
+    raise AttributeError("%s has no attribute %s", __name__, name)
